logging: read the full 64-bit rights fields of fdstat
fd_fdstat_get logs fs_rights_base and fs_rights_inheriting, and both were
read with le.Uint16 in internal/wasip1/logging/logging.go. Both fields are
64-bit: fdFdstatGetFn writes them with le.PutUint64 at offsets 8 and 16 in
imports/wasi_snapshot_preview1/fs.go writeFdstat. Every right at or above
bit 16 was therefore dropped from the log line, so a regular file logged
fs_rights_base without FD_FILESTAT_GET, FD_FILESTAT_SET_SIZE,
FD_FILESTAT_SET_TIMES or POLL_FD_READWRITE even though fileRightsBase sets
all four, and a directory lost everything from PATH_RENAME_SOURCE onward.
The path_open parameter logger reads the same rights straight off the stack
slot and prints them in full, so the two log lines disagreed about the same
values.

This reads both fields with le.Uint64 and updates the fdstat log lines
asserted in imports/wasi_snapshot_preview1/fs_test.go and
cmd/wazero/wazero_test.go, which now match the fs_rights_base bytes those
same test cases already assert in expectedMemory.

+++ internal/wasip1/logging/logging.go
@@ -274,6 +274,10 @@ func (i logFilestat) Log(_ context.Context, mod api.Module, w logging.Writer, pa
 
 type logFdstat uint32
 
+// Log writes the fdstat fields: fs_filetype is a u8 at offset 0, fs_flags a
+// u16 at offset 2, and fs_rights_base and fs_rights_inheriting are both the
+// 64-bit rights type, at offsets 8 and 16.
+// See https://github.com/WebAssembly/WASI/blob/snapshot-01/phases/snapshot/docs.md#-fdstat-struct
 func (i logFdstat) Log(_ context.Context, mod api.Module, w logging.Writer, params []uint64) {
 	offset, byteCount := uint32(params[i]), uint32(24)
 	if buf, ok := mod.Memory().Read(offset, byteCount); ok {
@@ -282,9 +286,9 @@ func (i logFdstat) Log(_ context.Context, mod api.Module, w logging.Writer, para
 		w.WriteString(",fdflags=")                            //nolint
 		w.WriteString(FdFlagsString(int(le.Uint16(buf[2:])))) //nolint
 		w.WriteString(",fs_rights_base=")                     //nolint
-		w.WriteString(RightsString(int(le.Uint16(buf[8:]))))  //nolint
+		w.WriteString(RightsString(int(le.Uint64(buf[8:]))))  //nolint
 		w.WriteString(",fs_rights_inheriting=")               //nolint
-		w.WriteString(RightsString(int(le.Uint16(buf[16:])))) //nolint
+		w.WriteString(RightsString(int(le.Uint64(buf[16:])))) //nolint
 		w.WriteString("}")                                    //nolint
 	}
 }
